Refactored Garbage Collector
Besides renaming it to MemoryManager, the code was cleaned up a bit.
Cells are now allocated directly with a call to calloc. The reaason for
this change is that previously the declaration for a MemoryCell
(previously called just Cell) contained a lot of stuff which fight
against the C++ compiler. This was really confusing and ugly.