Fixed buffer overflow
Adding a string to a buffer which is too small to hold it will cause it
to allocate more memory for itself to actually hold the string. I
incorrectly assumed doubling the size of the buffer was sufficient,
which it obviously isn't. This commit fixes the bug (which I've spent
some hours on) and also changes some other stuff all over the codebase
which I've lost track off.