WIP: Omit the argument field if not used
There is still a bug in the code which I can't figure out.
All the evidence points to the GC freeing cells it shouldn't.

The bug can be reproduced by running the runtime-profiler.ch.
It should fail at line (check disasm output) 50.
The reason is that the lambda is being freed before the call can happen.

This is most likely some bug with temporary nodes which are freed before
VM internal code can use them. Using a ManagedContext didn't help.

Further debugging needs a smaller program which produces the same error.

Note: Configuring the GC to have a single cell at startup, causes a
segfault as well, so maybe we are missing some ManagedContexts in some
other places of the VM.