Implemented span table in HeapRegion
Each HeapRegion gets divided into a set of spans. Each span stores the last known object offset inside that span and a dirty flag.

When a mutator thread writes a reference to an object in an eden region to an object inside an old region, it will set the dirty flag in the corresponding span in the old region. (To be implemented in future commit)