Implement compacting generational GC
Still got some bugs to work out, such as marking of function call arguments.

Currently regular call arguments are stored on the stack and are correctly tracked and updated by the GC, however for initial fiber calls, the arguments are stored in a tuple. That tuple gets invalidated during a GC and with it the arguments pointer.