GC correctly mark and move frame arguments
frame.arguments points to a memory region containing argc RawValue elements. This memory region can either be in a frame stack value buffer or inside a heap tuple.

If the region points to a tuple, frame.argument_tuple is a reference to that tuple. if the GC moves the tuple containing the memory for the arguments, the frame.arguments pointer needs to be moved to the new location of the tuple.