GC optimize heap memory sweep
Profiling suggested that the main cost of the GC collection came from having to traverse all eden regions in search of dead objects. By storing a pointer to objects that might contain external heap pointers during allocation, the GC only needs to traverse through this list of objects and check them, massively reducing the amount of objects we need to check.

Results in a 92% reduction in average GC cycle duration on a GC stress-test benchmark.